pyproject.toml
[project]
name = "smart-search"
version = "0.1.0"
description = "AI-powered search with zero hallucinations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.8.1",
    "cohere>=5.17.0",
    "ddgs>=9.5.5",
    "dspy>=3.0.3",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.116.1",
    "lancedb>=0.25.0",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "reflex>=0.6.0",
    "requests>=2.31.0",
    "tantivy>=0.24.0",
    "uvicorn>=0.35.0",
]

[project.scripts]
smart-search = "cli.smart_search:main"
